mcp-server-webhook-helper/pyproject.toml
[project]
name = "mcp-server-webhook-helper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.12.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
